+++ Connection.php
@@ -50,6 +50,9 @@ class Connection extends Component
      *  //'auth' => ['username' => 'yiiuser', 'password' => 'yiipw'], // Disabled auth regardless of `auth` property of the class
      * ]
      * ```
+     *
+     *  - `protocol`: explicitly sets the protocol for the current node (useful when manually defining a HTTPS cluster)
+     *
      * @see http://www.elastic.co/guide/en/elasticsearch/reference/current/cluster-nodes-info.html#cluster-nodes-info
      */
     public $nodes = [
@@ -72,6 +75,15 @@ class Connection extends Component
      */
     public $auth = [];
     /**
+     * Elasticsearch has no knowledge of protocol used to access its nodes. Specifically, cluster autodetection request
+     * returns node hosts and ports, but not the protocols to access them. Therefore we need to specify a default protocol here,
+     * which can be overridden for specific nodes in the [[nodes]] property.
+     * If [[autodetectCluster]] is true, all nodes received from cluster will be set to use the protocol defined by [[defaultProtocol]]
+     * @var string Default protocol to connect to nodes
+     * @since 2.0.5
+     */
+    public $defaultProtocol = 'http';
+    /**
      * @var float timeout to use for connecting to an elasticsearch node.
      * This value will be used to configure the curl `CURLOPT_CONNECTTIMEOUT` option.
      * If not set, no explicit timeout will be set for curl.
@@ -92,9 +104,15 @@ class Connection extends Component
 
     public function init()
     {
-        foreach ($this->nodes as $node) {
+        foreach ($this->nodes as &$node) {
             if (!isset($node['http_address'])) {
                 throw new InvalidConfigException('Elasticsearch node needs at least a http_address configured.');
+            }
+            if (!isset($node['protocol'])) {
+                $node['protocol'] = $this->defaultProtocol;
+            }
+            if (!in_array($node['protocol'], ['http', 'https'])) {
+                throw new InvalidConfigException('Valid node protocol settings are "http" and "https".');
             }
         }
     }
@@ -151,22 +169,35 @@ class Connection extends Component
     {
         $node = reset($this->nodes);
         $host = $node['http_address'];
+        $protocol = isset($node['protocol']) ? $node['protocol'] : $this->defaultProtocol;
         if (strncmp($host, 'inet[/', 6) === 0) {
             $host = substr($host, 6, -1);
         }
-        $response = $this->httpRequest('GET', 'http://' . $host . '/_nodes');
+        $response = $this->httpRequest('GET', "$protocol://$host/_nodes/_all/http");
         if (!empty($response['nodes'])) {
+            $nodes = $response['nodes'];
+        } else {
+            $nodes = [];
+        }
+
+        foreach ($nodes as $key => &$node) {
             // Make sure that nodes have an 'http_address' property, which is not the case if you're using AWS
-            // Elasticsearch service (at least as of Oct., 2015).
-            foreach ($response['nodes'] as &$node) {
-                if (!isset($node['http_address'])) {
-                    $node['http_address'] = $host;
-                }
+            // Elasticsearch service (at least as of Oct., 2015). - TO BE VERIFIED
+            // Temporary workaround - simply ignore all invalid nodes
+            if (!isset($node['http']['publish_address'])) {
+                unset($nodes[$key]);
             }
-            $this->nodes = $response['nodes'];
+            $node['http_address'] = $node['http']['publish_address'];
+
+            //Protocol is not a standard ES node property, so we add it manually
+            $node['protocol'] = $this->defaultProtocol;
+        }
+
+        if (!empty($nodes)) {
+            $this->nodes = array_values($nodes);
         } else {
             curl_close($this->_curl);
-            throw new Exception('Cluster autodetection did not find any active node.');
+            throw new Exception('Cluster autodetection did not find any active nodes.');
         }
     }
 
@@ -189,7 +220,7 @@ class Connection extends Component
             return;
         }
         Yii::trace('Closing connection to elasticsearch. Active node was: '
-            . $this->nodes[$this->activeNode]['http_address'], __CLASS__);
+            . $this->nodes[$this->activeNode]['http']['publish_address'], __CLASS__);
         $this->activeNode = null;
         if ($this->_curl) {
             curl_close($this->_curl);
@@ -361,7 +392,11 @@ class Connection extends Component
             }
         }
 
-        return [$this->nodes[$this->activeNode]['http_address'], $url];
+        $node = $this->nodes[$this->activeNode];
+        $protocol = isset($node['protocol']) ? $node['protocol'] : $this->defaultProtocol;
+        $host = $node['http_address'];
+
+        return [$protocol, $host, $url];
     }
 
     /**
@@ -440,15 +475,15 @@ class Connection extends Component
         }
 
         if (is_array($url)) {
-            list($host, $q) = $url;
+            list($protocol, $host, $q) = $url;
             if (strncmp($host, 'inet[', 5) == 0) {
                 $host = substr($host, 5, -1);
                 if (($pos = strpos($host, '/')) !== false) {
                     $host = substr($host, $pos + 1);
                 }
             }
-            $profile = $method . ' ' . $q . '#' . $requestBody;
-            $url = 'http://' . $host . '/' . $q;
+            $profile = "$method $q#$requestBody";
+            $url = "$protocol://$host/$q";
         } else {
             $profile = false;
         }
@@ -506,7 +541,7 @@ class Connection extends Component
         } elseif ($responseCode == 404) {
             return false;
         } else {
-            throw new Exception("Elasticsearch request failed with code $responseCode.", [
+            throw new Exception("Elasticsearch request failed with code $responseCode. Response body:\n{$body}", [
                 'requestMethod' => $method,
                 'requestUrl' => $url,
                 'requestBody' => $requestBody,
@@ -525,7 +560,6 @@ class Connection extends Component
             CURLOPT_WRITEFUNCTION => null,
             CURLOPT_READFUNCTION => null,
             CURLOPT_PROGRESSFUNCTION => null,
-            CURLOPT_POST => false,
             CURLOPT_POSTFIELDS => null,
         ];
         curl_setopt_array($this->_curl, $unsetValues);