+++ tests/QueryBuilderTest.php
@@ -51,8 +51,27 @@ class QueryBuilderTest extends TestCase
         $build = $queryBuilder->build($query);
         $this->assertTrue(array_key_exists('queryParts', $build));
         $this->assertTrue(array_key_exists('query', $build['queryParts']));
-        $this->assertFalse(array_key_exists('match_all', $build['queryParts']), 'Match all should not be set');
         $this->assertSame($queryParts, $build['queryParts']['query']);
+        $this->assertFalse(array_key_exists('match_all', $build['queryParts']), 'Match all should not be set');
+    }
+
+    /**
+     * @group postfilter
+     */
+    public function testQueryBuilderPostFilterQuery()
+    {
+        $postFilter = [
+            'bool' => [
+                'must' => [
+                    ['term' => ['title' => 'yii test']]
+                ]
+            ]
+        ];
+        $queryBuilder = new QueryBuilder($this->getConnection());
+        $query = new Query();
+        $query->postFilter($postFilter);
+        $build = $queryBuilder->build($query);
+        $this->assertSame($postFilter, $build['queryParts']['post_filter']);
     }
 
     public function testYiiCanBeFoundByQuery()
@@ -72,9 +91,11 @@ class QueryBuilderTest extends TestCase
                 'boost_mode' => 'replace',
                 'query' => ['term' => ['title' => 'yii']],
                 'functions' => [
-                    ['script_score' => [
-                        'script' => "doc['weight'].getValue()",
-                    ]],
+                    [
+                        'script_score' => [
+                            'script' => "doc['weight'].getValue()",
+                        ]
+                    ],
                 ],
             ],
         ];
@@ -118,59 +139,128 @@ class QueryBuilderTest extends TestCase
     {
         // >= 2010-01-15, 3 results
         $result = (new Query())
-                    ->from('yiitest', 'article')
-                    ->where(['>=', 'created_at', '2010-01-15'])
-                    ->search($this->getConnection());
+            ->from('yiitest', 'article')
+            ->where(['>=', 'created_at', '2010-01-15'])
+            ->search($this->getConnection());
         $this->assertEquals(3, $result['hits']['total']);
 
         // >= 2010-01-15, 3 results
         $result = (new Query())
-                    ->from('yiitest', 'article')
-                    ->where(['gte', 'created_at', '2010-01-15'])
-                    ->search($this->getConnection());
+            ->from('yiitest', 'article')
+            ->where(['gte', 'created_at', '2010-01-15'])
+            ->search($this->getConnection());
         $this->assertEquals(3, $result['hits']['total']);
 
         // > 2010-01-15, 2 results
         $result = (new Query())
-                    ->from('yiitest', 'article')
-                    ->where(['>', 'created_at', '2010-01-15'])
-                    ->search($this->getConnection());
+            ->from('yiitest', 'article')
+            ->where(['>', 'created_at', '2010-01-15'])
+            ->search($this->getConnection());
         $this->assertEquals(2, $result['hits']['total']);
 
         // > 2010-01-15, 2 results
         $result = (new Query())
-                    ->from('yiitest', 'article')
-                    ->where(['gt', 'created_at', '2010-01-15'])
-                    ->search($this->getConnection());
+            ->from('yiitest', 'article')
+            ->where(['gt', 'created_at', '2010-01-15'])
+            ->search($this->getConnection());
         $this->assertEquals(2, $result['hits']['total']);
 
         // <= 2010-01-20, 3 results
         $result = (new Query())
-                    ->from('yiitest', 'article')
-                    ->where(['<=', 'created_at', '2010-01-20'])
-                    ->search($this->getConnection());
+            ->from('yiitest', 'article')
+            ->where(['<=', 'created_at', '2010-01-20'])
+            ->search($this->getConnection());
         $this->assertEquals(3, $result['hits']['total']);
 
         // <= 2010-01-20, 3 results
         $result = (new Query())
-                    ->from('yiitest', 'article')
-                    ->where(['lte', 'created_at', '2010-01-20'])
-                    ->search($this->getConnection());
+            ->from('yiitest', 'article')
+            ->where(['lte', 'created_at', '2010-01-20'])
+            ->search($this->getConnection());
         $this->assertEquals(3, $result['hits']['total']);
 
         // < 2010-01-20, 2 results
         $result = (new Query())
-                    ->from('yiitest', 'article')
-                    ->where(['<', 'created_at', '2010-01-20'])
-                    ->search($this->getConnection());
+            ->from('yiitest', 'article')
+            ->where(['<', 'created_at', '2010-01-20'])
+            ->search($this->getConnection());
         $this->assertEquals(2, $result['hits']['total']);
 
         // < 2010-01-20, 2 results
         $result = (new Query())
-                    ->from('yiitest', 'article')
-                    ->where(['lt', 'created_at', '2010-01-20'])
-                    ->search($this->getConnection());
+            ->from('yiitest', 'article')
+            ->where(['lt', 'created_at', '2010-01-20'])
+            ->search($this->getConnection());
         $this->assertEquals(2, $result['hits']['total']);
     }
 
+    public function testNotCondition()
+    {
+        $titles = [
+            'Symfony2 is another framework',
+            'yii test',
+            'nonexistent',
+        ];
+        $result = (new Query)
+            ->from('yiitest', 'article')
+            ->where([ 'not', [ 'in', 'title.keyword', $titles ] ])
+            ->search($this->getConnection());
+        $this->assertEquals(2, $result['hits']['total']);
+    }
+
+    public function testInCondition()
+    {
+        $titles = [
+            'Symfony2 is another framework',
+            'yii test',
+            'nonexistent',
+        ];
+        $result = (new Query)
+            ->from('yiitest', 'article')
+            ->where([ 'in', 'title.keyword', $titles ])
+            ->search($this->getConnection());
+        $this->assertEquals(2, $result['hits']['total']);
+    }
+
+    public function testBuildNotCondition()
+    {
+        $db = $this->getConnection();
+        $qb = new QueryBuilder($db);
+
+        $cond = [ 'title' => 'xyz' ];
+        $operands = [ $cond ];
+
+        $expected = [
+            'bool' => [
+                'must_not' => [
+                    'bool' => [ 'must' => [ ['term'=>['title'=>'xyz']] ] ],
+                ],
+            ]
+        ];
+        $result = $this->invokeMethod($qb, 'buildNotCondition', ['not',$operands]);
+        $this->assertEquals($expected, $result);
+    }
+
+    public function testBuildInCondition()
+    {
+        $db = $this->getConnection();
+        $qb = new QueryBuilder($db);
+
+        $expected = [
+            'terms' => ['foo' => ['bar1', 'bar2']],
+        ];
+        $result = $this->invokeMethod($qb, 'buildInCondition', [
+            'in',
+            ['foo',['bar1','bar2']]
+        ]);
+        $this->assertEquals($expected, $result);
+    }
+
+    public function invokeMethod($obj, $methodName, $args)
+    {
+        $reflection = new \ReflectionObject($obj);
+        $method = $reflection->getMethod($methodName);
+        $method->setAccessible(true);
+        return $method->invokeArgs($obj, $args);
+    }
 }